Potential fix for code scanning alert no. 15: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger type (double) to avoid overflow. This can be achieved by casting one of the operands to double before performing the multiplication. This way, the multiplication will be done in double precision, preventing any overflow that might occur with float precision.

Specifically, we will cast lCurrentValue to double before the multiplication on line 563. This change will ensure that the multiplication is performed using double precision, and the result will be correctly accumulated in the double variable lNorms[i].

+++ libopenjpeg-2.3.1/src/lib/openjp2/mct.c
@@ -560,7 +560,7 @@ void opj_calculate_norms(OPJ_FLOAT64 * pNorms,
         for (j = 0; j < pNbComps; ++j) {
             lCurrentValue = lMatrix[lIndex];
             lIndex += pNbComps;
-            lNorms[i] += lCurrentValue * lCurrentValue;
+            lNorms[i] += (double)lCurrentValue * lCurrentValue;
         }
         lNorms[i] = sqrt(lNorms[i]);
     }
